requirements.txt
opencv-python
mediapipe==0.10.9
pillow
numpy